[BACKLOG-18687] DET - contextual menu doesn't appear for numeric types in categorical axis
