+++ impl/client/src/main/javascript/web/pentaho/ccc/visual/axes/DiscreteAxis.js
@@ -17,38 +17,12 @@ define([
   "cdf/lib/CCC/def",
   "cdf/lib/CCC/pvc",
   "./AbstractAxis",
-  "../_util",
   "pentaho/data/filter/isEqual"
-], function(def, pvc, AbstractAxis, util, isEqualFactory) {
+], function(def, pvc, AbstractAxis, isEqualFactory) {
 
   "use strict";
 
   return AbstractAxis.extend({
-    _nonMultiMappingAttrInfoFilter: function(maInfo) {
-      return maInfo.role !== this.chart._multiRole;
-    },
-
-    _isNullMember: function(complex, maInfo) {
-      var atom = complex.atoms[maInfo.cccDimName];
-      return util.isNullMember(atom.value);
-    },
-
-    _buildMappingAttrInfoHtmlTooltip: function(lines, complex, context, maInfo, index) {
-      /*
-       * Multi-chart formulas are not shown in the tooltip
-       * They're on the small chart's title.
-       *
-       * Also, if the chart hides null members,
-       * don't show them in the tooltip.
-       * Using the scene's group, preferably, because the datum (here the complex) may have dimensions
-       * that are null in the groups' own atoms.
-       */
-      if(this._nonMultiMappingAttrInfoFilter(maInfo) &&
-         !(this.chart._hideNullMembers && this._isNullMember(context.scene.group || complex, maInfo))) {
-        this.base.apply(this, arguments);
-      }
-    },
-
     complexToFilter: function(complex) {
       var filter = null;
 
@@ -62,7 +36,10 @@ define([
         if(value != null) {
           if(!IsEqual) IsEqual = context.get(isEqualFactory);
 
-          var operand = new IsEqual({property: maInfo.attr.name, value: {_: "string", v: value, f: atom.label}});
+          var attrType = maInfo.attr.type;
+          var valueType = attrType === "number" ? attrType : "string";
+
+          var operand = new IsEqual({property: maInfo.attr.name, value: {_: valueType, v: value, f: atom.label}});
           filter = filter ? filter.and(operand) : operand;
         }
       });
